refactor(.github/actions) make file naming consistent
